docker: update to 18.09.3

+++ packages/addons/service/docker/package.mk
@@ -1,33 +1,17 @@
-################################################################################
-#      This file is part of LibreELEC - https://libreelec.tv
-#      Copyright (C) 2009-2017 Lukas Rusak (lrusak@libreelec.tv)
-#      Copyright (C) 2017-present Team LibreELEC
-#
-#  LibreELEC is free software: you can redistribute it and/or modify
-#  it under the terms of the GNU General Public License as published by
-#  the Free Software Foundation, either version 2 of the License, or
-#  (at your option) any later version.
-#
-#  LibreELEC is distributed in the hope that it will be useful,
-#  but WITHOUT ANY WARRANTY; without even the implied warranty of
-#  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#  GNU General Public License for more details.
-#
-#  You should have received a copy of the GNU General Public License
-#  along with LibreELEC.  If not, see <http://www.gnu.org/licenses/>.
-################################################################################
+# SPDX-License-Identifier: GPL-2.0-or-later
+# Copyright (C) 2009-2017 Lukas Rusak (lrusak@libreelec.tv)
+# Copyright (C) 2017-present Team LibreELEC (https://libreelec.tv)
 
 PKG_NAME="docker"
-PKG_VERSION="17.10.0"
-PKG_SHA256="90f54b988d5241ee0472800e139e0628ae8a58dac168bb32fdb031383f3b46be"
-PKG_REV="117"
+PKG_VERSION="18.09.3"
+PKG_REV="122"
 PKG_ARCH="any"
-PKG_ADDON_PROJECTS="Generic RPi RPi2 WeTek_Hub WeTek_Play_2 Odroid_C2"
+PKG_ADDON_PROJECTS="Generic RPi RPi2"
 PKG_LICENSE="ASL"
 PKG_SITE="http://www.docker.com/"
-PKG_URL="https://github.com/docker/docker-ce/archive/v${PKG_VERSION}-ce.tar.gz"
-PKG_SOURCE_DIR="docker-ce-${PKG_VERSION}-ce"
-PKG_DEPENDS_TARGET="toolchain sqlite go:host containerd runc libnetwork tini"
+PKG_URL="https://github.com/docker/docker-ce/archive/v${PKG_VERSION}.tar.gz"
+PKG_SOURCE_DIR="docker-ce-$PKG_VERSION"
+PKG_DEPENDS_TARGET="toolchain sqlite go:host containerd runc libnetwork tini systemd"
 PKG_SECTION="service/system"
 PKG_SHORTDESC="Docker is an open-source engine that automates the deployment of any application as a lightweight, portable, self-sufficient container that will run virtually anywhere."
 PKG_LONGDESC="Docker containers can encapsulate any payload, and will run consistently on and between virtually any server. The same container that a developer builds and tests on a laptop will run at scale, in production*, on VMs, bare-metal servers, OpenStack clusters, public instances, or combinations of the above."
@@ -40,6 +24,7 @@ PKG_ADDON_TYPE="xbmc.service"
 configure_target() {
   export DOCKER_BUILDTAGS="daemon \
                            autogen \
+                           exclude_graphdriver_btrfs \
                            exclude_graphdriver_devicemapper \
                            exclude_graphdriver_aufs \
                            journald"
@@ -55,9 +40,9 @@ configure_target() {
         arm1176jzf-s)
           export GOARM=6
           ;;
-        cortex-a7|cortex-a9)
-         export GOARM=7
-         ;;
+        *)
+          export GOARM=7
+          ;;
       esac
       ;;
     aarch64)
@@ -71,26 +56,56 @@ configure_target() {
   export CGO_CFLAGS=$CFLAGS
   export LDFLAGS="-w -linkmode external -extldflags -Wl,--unresolved-symbols=ignore-in-shared-libs -extld $CC"
   export GOLANG=$TOOLCHAIN/lib/golang/bin/go
-  export GOPATH=$PKG_BUILD/.gopath:$PKG_BUILD/.gopath_cli
+  export GOPATH=$PKG_BUILD/.gopath_cli:$PKG_BUILD/.gopath
   export GOROOT=$TOOLCHAIN/lib/golang
   export PATH=$PATH:$GOROOT/bin
 
   mkdir -p $PKG_BUILD/.gopath
   mkdir -p $PKG_BUILD/.gopath_cli
+
   PKG_ENGINE_PATH=$PKG_BUILD/components/engine
   PKG_CLI_PATH=$PKG_BUILD/components/cli
+
   if [ -d $PKG_ENGINE_PATH/vendor ]; then
     mv $PKG_ENGINE_PATH/vendor $PKG_BUILD/.gopath/src
   fi
+
   if [ -d $PKG_CLI_PATH/vendor ]; then
     mv $PKG_CLI_PATH/vendor $PKG_BUILD/.gopath_cli/src
   fi
-  ln -fs $PKG_ENGINE_PATH $PKG_BUILD/.gopath/src/github.com/docker/docker
-  ln -fs $PKG_CLI_PATH $PKG_BUILD/.gopath_cli/src/github.com/docker/cli
+
+  # Fix missing/incompatible .go files
+  cp -rf $PKG_BUILD/.gopath/src/github.com/moby/buildkit/frontend/* $PKG_BUILD/.gopath_cli/src/github.com/moby/buildkit/frontend
+  cp -rf $PKG_BUILD/.gopath/src/github.com/moby/buildkit/frontend/gateway/* $PKG_BUILD/.gopath_cli/src/github.com/moby/buildkit/frontend/gateway
+  cp -rf $PKG_BUILD/.gopath/src/github.com/moby/buildkit/solver/* $PKG_BUILD/.gopath_cli/src/github.com/moby/buildkit/solver
+  cp -rf $PKG_BUILD/.gopath/src/github.com/moby/buildkit/util/progress/* $PKG_BUILD/.gopath_cli/src/github.com/moby/buildkit/util/progress
+  cp -rf $PKG_BUILD/.gopath/src/github.com/docker/swarmkit/manager/* $PKG_BUILD/.gopath_cli/src/github.com/docker/swarmkit/manager
+  cp -rf $PKG_BUILD/.gopath/src/github.com/coreos/etcd/raft/* $PKG_BUILD/.gopath_cli/src/github.com/coreos/etcd/raft
+  cp -rf $PKG_BUILD/.gopath/src/golang.org/x/* $PKG_BUILD/.gopath_cli/src/golang.org/x
+  cp -rf $PKG_BUILD/.gopath/src/github.com/opencontainers/runtime-spec/specs-go/* $PKG_BUILD/.gopath_cli/src/github.com/opencontainers/runtime-spec/specs-go
+
+  rm -rf $PKG_BUILD/.gopath_cli/src/github.com/containerd/containerd
+  mkdir -p $PKG_BUILD/.gopath_cli/src/github.com/containerd/containerd
+  cp -rf $PKG_BUILD/.gopath/src/github.com/containerd/containerd/* $PKG_BUILD/.gopath_cli/src/github.com/containerd/containerd
+
+  rm -rf $PKG_BUILD/.gopath_cli/src/github.com/containerd/continuity
+  mkdir -p $PKG_BUILD/.gopath_cli/src/github.com/containerd/continuity
+  cp -rf $PKG_BUILD/.gopath/src/github.com/containerd/continuity/* $PKG_BUILD/.gopath_cli/src/github.com/containerd/continuity
+
+  mkdir -p $PKG_BUILD/.gopath_cli/src/github.com/docker/docker/builder
+  cp -rf $PKG_ENGINE_PATH/builder/* $PKG_BUILD/.gopath_cli/src/github.com/docker/docker/builder
+
+  if [ ! -L $PKG_BUILD/.gopath/src/github.com/docker/docker ];then
+    ln -fs $PKG_ENGINE_PATH $PKG_BUILD/.gopath/src/github.com/docker/docker
+  fi
+
+  if [ ! -L $PKG_BUILD/.gopath_cli/src/github.com/docker/cli ];then
+    ln -fs $PKG_CLI_PATH $PKG_BUILD/.gopath_cli/src/github.com/docker/cli
+  fi
 
   # used for docker version
-  export GITCOMMIT=${PKG_VERSION}-ce
-  export VERSION=${PKG_VERSION}-ce
+  export GITCOMMIT=${PKG_VERSION}
+  export VERSION=${PKG_VERSION}
   export BUILDTIME="$(date --utc)"
 
   cd $PKG_ENGINE_PATH
@@ -117,14 +132,14 @@ addon() {
     cp -P $PKG_BUILD/bin/dockerd $ADDON_BUILD/$PKG_ADDON_ID/bin
 
     # containerd
-    cp -P $(get_build_dir containerd)/bin/containerd $ADDON_BUILD/$PKG_ADDON_ID/bin/docker-containerd
-    cp -P $(get_build_dir containerd)/bin/containerd-shim $ADDON_BUILD/$PKG_ADDON_ID/bin/docker-containerd-shim
+    cp -P $(get_build_dir containerd)/bin/containerd $ADDON_BUILD/$PKG_ADDON_ID/bin/containerd
+    cp -P $(get_build_dir containerd)/bin/containerd-shim $ADDON_BUILD/$PKG_ADDON_ID/bin/containerd-shim
 
     # libnetwork
     cp -P $(get_build_dir libnetwork)/bin/docker-proxy $ADDON_BUILD/$PKG_ADDON_ID/bin/docker-proxy
 
     # runc
-    cp -P $(get_build_dir runc)/bin/runc $ADDON_BUILD/$PKG_ADDON_ID/bin/docker-runc
+    cp -P $(get_build_dir runc)/bin/runc $ADDON_BUILD/$PKG_ADDON_ID/bin/runc
 
     # tini
     cp -P $(get_build_dir tini)/.$TARGET_NAME/tini-static $ADDON_BUILD/$PKG_ADDON_ID/bin/docker-init
